Update Api.php
Can you please add these changes so that Clickatell can provide us with a from number in cases where we have to use the "mo" parameter.  And also the "cliMsgId" parameter so we can assign every sms a unique key so when Clickatell does the callback it is easier to find the specific sms.

Or alternatively can you please update the code so that when we use the "request" event and inject additional parameters, it persists through to the final request. At this moment the "request" event provided can only be used for logging or debugging since all changes are ignored before the request is sent.